requirements.txt
Django==4.2.5
django-split-settings==1.2.0
django-ninja==0.22.2
daphne==4.0.0